folder reorg, model evaluation and chart generation
