--- index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<script src="http://d3js.org/d3.v3.min.js"></script>
-<script src="d3.layout.cloud.js"></script>
-
-<head>
-    <title>Word Cloud Example</title>
-</head>
-
-<body>
-<script src="cloud-controller.js"></script>
-</body>
-
-
-
-
-</html>
